Language.StructuredScript.Parsers

Description

function to define the grammer of the scheme

Documentation

data Expr

SST Grammer

Useage examples:

x := true;

|

Expression Grammar

Constructors

Var String

Con Const

Uno Unop Expr

Duo Duop Expr Expr

type VType = Const

newtype VarTable

Constructors

VT (HashMap Text VType)

emptyVTable :: VarTable

data Const

Define basic Data Types | Boolean | Integer | Double | String | Character

Constructors

ConstBool Bool

ConstInteger Integer

ConstString String

ConstChar Char

ConstDouble Double

data Unop

Unary Operators [~,-]

Constructors

Not

Neg

data Duop

Binary Operators | Relational Operators -> [&&, ||, XOR] | Logical Operators -> [>, ==,=, <] | Arithmetic Operators -> [+, * , /, -, MOD]

Constructors

And

Or

XOr

IsSet

Greater

Less

Equal

GreaterEqual

LessEqual

NotEqual

Add

Mul

Div

Sub

Mod

Concat

Pow

data Stmt

Statement Grammar | ~stmt | External | x: = a + b | if (a == b) stmt1 stmt2

Constructors

Nop

External

String := Expr

If Expr Stmt Stmt

Seq [Stmt]

def :: GenLanguageDef String st Identity

General Language Rules | Comment: "* * | Reserved Words: True | False | nop | if | then | else |end_if

sst_lexer :: GenTokenParser String u Identity

Create specific parsers from the generic parser creator in Parsec

sst_parens :: ParsecT String u Identity a -> ParsecT String u Identity a

Lex Parentheses

sst_identifier :: ParsecT String u Identity String

Lex Indentifiers

sst_reservedOp :: String -> ParsecT String u Identity ()

Lex Reserved Operators

sst_reserved :: String -> ParsecT String u Identity ()

Lex Reserved Keywords

sst_semiSep1 :: ParsecT String u Identity a -> ParsecT String u Identity [a]

Lex Semicolon

sst_natural :: ParsecT String u Identity Integer

Numeric Literals

sst_integer :: ParsecT String u Identity Integer

sst_decimal :: ParsecT String u Identity Integer

sst_octal :: ParsecT String u Identity Integer

sst_hexadecimal :: ParsecT String u Identity Integer

sst_double :: ParsecT String u Identity Double

sst_naturalOrDouble :: ParsecT String u Identity (Either Integer Double)

sst_stringLiteral :: ParsecT String u Identity String

String Literal

sst_charLiteral :: ParsecT String u Identity Char

sst_whiteSpace :: ParsecT String u Identity ()

White Space

evalStmt :: VarTable -> Stmt -> Either String VarTable

insertToLut :: VarTable -> Stmt -> Either String VarTable

exprparser :: Parser Expr

table :: [[Operator String u Identity Expr]]

Operators Syntax Table Construction

term :: ParsecT String () Identity Expr

Term parse each parser one by one

evalExpr :: VarTable -> Expr -> Either String Const

Evaluate Expression

unopLookUp :: Unop -> Const -> Either String Const

Unary Operators | Not Operator -> [~, NOT]

duopLookUp :: Duop -> Const -> Const -> Either String Const

Binary Operators ==Arithmetic Operators== | Addition Operator -> [ + ]

boolTParser :: ParsecT String t Identity Expr

Parse Boolean

boolFParser :: ParsecT String t Identity Expr

intParser :: ParsecT String u Identity Expr

Parses any integer type constant to an Integer

naturalOrDoubleParser :: ParsecT String u Identity Expr

Parses either an integer or a Double into a Integer or Double

stringParser :: ParsecT String u Identity Expr

Parses any string type constant to an String

charParser :: ParsecT String u Identity Expr

Parses any char type constant to an Char

mainparser :: Parser Stmt

Main parser to intergrate the different parser and evluation functions

testString :: String

testList :: [Const]

sstParse :: String -> Either String Stmt

- | Takes a string and return an IO Type output play :: String -> IO () play inp = case parse mainparser "" inp of { Left err -> print err ; Right ans -> print ans }

- | Takes a string and returns an Stmt Type output run :: String -> Stmt run str = case parse mainparser "" str of Left e -> error $ show e Right r -> r |

Intergration for Parse, return an Either String or Stmt Type

sstEval :: VarTable -> Stmt -> Either String VarTable

Intergration for Evaluation of syntax, return an Either String or VarTable Type

sstLookupOutput :: VarTable -> Either String Const

Lookup output parameter from VarTable, return Either String or Const Type

sstInsertInput :: [Const] -> Either String VarTable

Test function sst, parse everything and returns an output sst = sstLookupOutput =<< (evalStmt emptyVTable =<< sstParse testString)

Insert input into the script from a list of data

sstInsertInputVector :: Vector Const -> Either String VarTable

sstTest :: [Const] -> String -> Either String Const

Intergration Test Function for inserting a list of data and parse into a certain output

sstTest2 = do vt <- sstInsertInput testList sstParse testString >>= sstEval vt >>= sstLookupOutput |